[SAFE-LD] Move unit tests from regression/ to unit/ld-frontend/
Rewrites the three standalone assert-based test suites (parser,
property, ir_gen) as Catch2 TEST_CASE / REQUIRE tests and registers
them via new_unit_test() in unit/ld-frontend/, consistent with how
all other frontends (Python, Jimple, clang-c) organise unit tests.

Fixture files moved from regression/ld/unit/*/fixtures/ to the shared
unit/ld-frontend/fixtures/ directory.  FIXTURE_DIR compile definition
points to CMAKE_CURRENT_SOURCE_DIR/fixtures at configure time.

regression/ld/CMakeLists.txt now contains only the ENABLE_LD_FRONTEND
guard; all unit test subdirectories are removed.
