+++ unit/ld-frontend/fixtures/coil_kind_attr.ld
