+++ unit/ld-frontend/fixtures/derived_type.ld
