+++ unit/ld-frontend/fixtures/mutual_exclusion_one_var.yaml
