+++ unit/ld-frontend/fixtures/valid_props.yaml
